Starexec mergesat

+++ scripts/make-starexec.sh
@@ -27,11 +27,39 @@ then
 	exit 1
 fi
 
+# shall we leave the directory behind?
+NO_CLEANUP=${NO_STP_CLEANUP:-}
+
+CMSCOMMIT=""
+MERGESATCOMMIT=""
+MINISATCOMMIT=""
+RISSCOMMIT=""
+# do we want to package Riss(for Coprocessor) or Sparrow as well?
+while getopts "c:g:m:r:" OPTION; do
+    case $OPTION in
+    c)
+        CMSCOMMIT="$OPTARG"
+        ;;
+    g)
+        MERGESATCOMMIT="$OPTARG"
+        ;;
+    m)
+        MINISATCOMMIT="$OPTARG"
+        ;;
+    r)
+        RISSCOMMIT="$OPTARG"
+        ;;
+    *)
+        echo "Unknown options provided"
+        ;;
+    esac
+done
+
 # give the package a unique name
 DESCRIPTION=$(git rev-parse --short HEAD)
 
 # make sure we clean up
-trap 'rm -rf $TMPD' EXIT
+[ -n "$NO_CLEANUP" ] || trap 'rm -rf $TMPD' EXIT
 TMPD=$(mktemp -d)
 
 # create the project directory
@@ -58,11 +86,12 @@ cd m4ri-20140914/
 make -j $(nproc)
 echo "Before proceeding, make sure you want to install m4ri version 20140914 from source"
 sleep 5
-sudo make install
+timeout 10 sudo make install || true
 cd ..
 
 git clone --depth 1 https://github.com/msoos/cryptominisat.git
 cd cryptominisat
+[ -z "$CMSCOMMIT" ] || git checkout -b stp-build "$CMSCOMMIT"
 echo "CryptoMinisat: $(git rev-parse --short HEAD)" >> ../COMMITS
 mkdir build
 cd build
@@ -74,6 +103,7 @@ cp cryptominisat/build/cryptominisat5 bin
 # Minisat
 git clone --depth 1 https://github.com/niklasso/minisat.git
 cd minisat
+[ -z "$MINISATCOMMIT" ] || git checkout -b stp-build "$MINISATCOMMIT"
 make r -j $(nproc)
 echo "Minisat: $(git rev-parse --short HEAD)" >> ../COMMITS
 cd ..
@@ -82,6 +112,7 @@ cp minisat/build/release/bin/minisat bin
 # Riss
 git clone --depth 1 https://github.com/conp-solutions/riss.git
 cd riss
+[ -z "$RISSCOMMIT" ] || git checkout -b stp-build "$RISSCOMMIT"
 echo "Riss: $(git rev-parse --short HEAD)" >> ../COMMITS
 mkdir -p build
 cd build
@@ -89,6 +120,15 @@ cmake .. -DDRATPROOF=OFF -DCMAKE_BUILD_TYPE=Release
 make riss-core -j $(nproc)
 cd ../..
 cp riss/build/bin/riss-core bin/riss
+
+# MergeSAT
+git clone --depth 1 https://github.com/conp-solutions/mergesat.git
+cd mergesat
+[ -z "$MERGESATCOMMIT" ] || git checkout -b stp-build "$MERGESATCOMMIT"
+echo "MergeSAT: $(git rev-parse --short HEAD)" >> ../COMMITS
+make r -j $(nproc)
+cd ..
+cp mergesat/build/release/bin/mergesat bin/mergesat
 
 # build STP
 mkdir -p stp-build
@@ -99,10 +139,10 @@ cmake -DSTATICCOMPILE:BOOL=ON \
  ../stp
 make -j $(nproc)
 cd ..
-cp stp-build/stp-2.1.2 bin
+cp stp-build/stp bin
 
 # Cleanup solver directories
-rm -rf stp-build minisat m4ri-20140914 cryptominisat riss stp ./*.tar.gz
+[ -n "$NO_CLEANUP" ] || rm -rf stp-build minisat m4ri-20140914 cryptominisat riss stp mergesat ./*.tar.gz
 
 # Get rid of extra symbols in solvers
 cd bin
